+++ services/index.js
@@ -4,7 +4,15 @@ const {
   getAllAnnonce,
   getAllColis,
 } = require("./userService/");
-const { createAddress, updateAnnonceById } = require("./annouceService/");
+const {
+  createAddress,
+  updateAnnonceById,
+  getAnnouce,
+  createContent,
+  createAllContent,
+  deleteContentByArray,
+} = require("./annouceService/");
+const { createImage, createAllImage } = require("./imageService/imageService");
 module.exports = {
   /* userService */
   handleRefreshToken,
@@ -12,8 +20,16 @@ module.exports = {
   getAllAnnonce,
   getAllColis,
   /* end userService */
+  /*imageService */
+  createImage,
+  createAllImage,
+  /*end imageService */
   /* annouceService */
   createAddress,
+  getAnnouce,
   updateAnnonceById,
+  createContent,
+  createAllContent,
+  deleteContentByArray,
   /* end annouceService */
 };
